+++ packages/terra-slide-panel/src/terra-dev-site/test/slide-panel/SlidePanelMedium.test.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import SlidePanel from '../../../SlidePanel';
+import SlidePanelMainContent from '../common/SlidePanelMainContent';
+import SlidePanelPanelContent from '../common/SlidePanelPanelContent';
+
+const SlidePanelLarge = () => (
+  <SlidePanel
+    mainContent={<SlidePanelMainContent />}
+    panelContent={<SlidePanelPanelContent />}
+    panelBehavior="squish"
+    panelPosition="start"
+    panelSize="medium"
+    isOpen
+  />
+);
+
+export default SlidePanelLarge;